requirements.txt
dnspython==2.6.1
Faker==24.3.0
mysql-connector-python==8.3.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
redis==5.0.3
six==1.16.0
